requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
SQLAlchemy[asyncio]==2.0.43
asyncpg==0.30.0
alembic==1.16.5
redis==6.4.0
celery==5.5.3
httpx==0.28.1
beautifulsoup4==4.13.5
bleach==6.2.0
PyYAML==6.0.2
transformers==4.56.0
torch>=2.9.0,<2.11.0
spacy==3.8.11
nltk==3.9.1
scikit-learn==1.7.1; python_version < "3.14"
prometheus-fastapi-instrumentator==7.1.0
structlog==25.4.0
slowapi==0.1.9
google-generativeai==0.8.5
tenacity==9.1.2
orjson==3.11.3
datasets==4.0.0
streamlit==1.49.1
